requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.6.0
marshmallow==3.21.1
Flask-Marshmallow==1.2.1
Flask-Limiter==3.8.0
redis==5.0.4
Flask-Caching==2.3.0
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.0.2